# src/docs/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Mercado Inclusivo",
    "version": "1.0.0",
    "description": "API para inclusão de PCDs no mercado de trabalho"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/auth/register": {
      "post": {
        "summary": "Registra um novo usuário",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Usuario"
              }
            }
          }
        },
        "responses": {
          "201": { "description": "Usuário criado com sucesso" },
          "400": { "description": "Usuário já existe" },
          "500": { "description": "Erro interno no servidor" }
        }
      }
    },
    "/auth/login": {
      "post": {
        "summary": "Faz login do usuário e retorna um token JWT",
        "tags": ["Autenticação"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": { "type": "string", "format": "email" },
                  "senha": { "type": "string" }
                },
                "required": ["email", "senha"]
              }
            }
          }
        },
        "responses": {
          "200": { "description": "Login realizado com sucesso (token JWT retornado)" },
          "401": { "description": "Credenciais inválidas" },
          "500": { "description": "Erro interno no servidor" }
        }
      }
    },
...
  },
...
  "components": {
    "schemas": {
      "Usuario": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "nome": { "type": "string" },
          "email": { "type": "string" },
          "senha": { "type": "string" },
          "tipo": { "type": "string", "enum": ["PCD", "empresa"] },
          "deficiencia": { "type": "string" }
        },
        "required": ["nome", "email", "senha", "tipo"]
      },
...
    }
  }
}